Subtle changes and corrected gitignore

+++ TP/include/demand.hpp
@@ -13,7 +13,7 @@ class Demand {
     private:
         // Atributos gerais
         int id;                 // Identificador
-        int time;               // Marcador de tempo de solicitação da demanda
+        double time;            // Marcador de tempo de solicitação da demanda
         Point2D origin;         // Ponto de origem
         Point2D destination;    // Ponto de destino
         State state;            // Estado da demanda: solicitada, individual, compartilhada ou completa
@@ -23,13 +23,13 @@ class Demand {
 
     public:
         // Construtores
-        Demand() : Demand(-1, -1, 0.0, 0.0, 0.0, 0.0) { };                  // Demanda "nula" (por definição)
-        Demand(int id, int t, double ox, double oy, double dx, double dy);  // Construtor completo
-        Demand(const Demand& other);                                              // Construtor de cópia
+        Demand() : Demand(-1, -1, 0.0, 0.0, 0.0, 0.0) { };                      // Demanda "nula" (por definição)
+        Demand(int id, double t, double ox, double oy, double dx, double dy);   // Construtor completo
+        Demand(const Demand& other);                                            // Construtor de cópia
 
         // Getters para acesso aos atributos
         int GetID();                                    // Retorna o id da demanda
-        int GetTime();                                  // Retorna o tempo de solicitação
+        double GetTime();                               // Retorna o tempo de solicitação
         Point2D& GetOrigin();                           // Retorna referência para o ponto de origem
         Point2D& GetDestination();                      // Retorna referência para o ponto de destino
         double OriginDistance(Demand& other);           // Calcula distância entre origens desta demanda e outra demanda
